fix for cross compile

+++ src/main/scala/xerial/sbt/Sonatype.scala
@@ -477,7 +477,7 @@ object Sonatype extends AutoPlugin with LogSupport {
     val eitherOp = for {
       client <- SonatypeCentralClient.fromCredentials(
         credentials,
-        readTimeoutMillis = extracted.get(sonatypeTimeoutMillis)
+        readTimeoutMillis = extracted.get(sonatypeTimeoutMillis).toLong
       )
       service = new SonatypeCentralService(client)
       res <-
